requirements.txt
numpy==1.26.4
sentence_transformers==3.4.1
torch==2.2.2
torch-geometric==2.6.1
neo4j==5.28.1
pandas==2.2.3
scikit-learn==1.6.1
transformers==4.49.0
umap-learn==0.5.7
hdbscan==0.8.40
matplotlib==3.10.1
seaborn==0.13.2
plotly==6.0.1
llama-cpp-python==0.3.8
typing==3.7.4.3

